Helper class to reduce client boilerplate

* Renamed BdfReader methods for consistency with BdfList/BdfDictionary
* Added readList() and readDictionary() methods to BdfReader
* Added ClientHelper to reduce boilerplate when converting messages and metadata to and from BDF
* Moved PrivateGroupFactory to the same package as ClientHelper


See merge request !114

+++ briar-api/src/org/briarproject/api/data/BdfWriter.java
@@ -11,8 +11,8 @@ public interface BdfWriter {
 
 	void writeNull() throws IOException;
 	void writeBoolean(boolean b) throws IOException;
-	void writeInteger(long l) throws IOException;
-	void writeFloat(double d) throws IOException;
+	void writeLong(long l) throws IOException;
+	void writeDouble(double d) throws IOException;
 	void writeString(String s) throws IOException;
 	void writeRaw(byte[] b) throws IOException;
